Extract preview view components
