+++ Preview/Views/InfoRow.swift
@@ -1,0 +1,19 @@
+import SwiftUI
+
+struct InfoRow: View {
+    let label: String
+    let value: String
+
+    var body: some View {
+        HStack(alignment: .top) {
+            Text(label)
+                .fontWeight(.medium)
+                .frame(minWidth: UIConstants.Size.minLabelWidth, alignment: .leading)
+                .foregroundColor(.secondary)
+
+            Text(value)
+                .textSelection(.enabled)
+                .frame(maxWidth: .infinity, alignment: .leading)
+        }
+    }
+}
