+++ Preview/Views/OverviewSection.swift
@@ -24,31 +24,3 @@ struct OverviewSection: View {
         }
     }
 }
-
-extension ExpirationStatus {
-    var color: Color {
-        switch self {
-        case .expired: .red
-        case .expiring: .orange
-        case .valid: UIConstants.Color.validGreen
-        }
-    }
-}
-
-struct InfoRow: View {
-    let label: String
-    let value: String
-
-    var body: some View {
-        HStack(alignment: .top) {
-            Text(label)
-                .fontWeight(.medium)
-                .frame(minWidth: UIConstants.Size.minLabelWidth, alignment: .leading)
-                .foregroundColor(.secondary)
-
-            Text(value)
-                .textSelection(.enabled)
-                .frame(maxWidth: .infinity, alignment: .leading)
-        }
-    }
-}
